backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
psycopg2-binary
alembic
supabase
pydantic
scikit-learn
pandas
joblib